+++ framework/packages/abstract-client/src/client.rs
@@ -28,9 +28,10 @@
 //! # Ok::<(), AbstractClientError>(())
 //! ```
 
-use abstract_core::objects::{namespace::Namespace, AccountId};
+use abstract_core::objects::{namespace::Namespace, salt::generate_instantiate_salt, AccountId};
 use abstract_interface::{
-    Abstract, AbstractAccount, AccountFactoryQueryFns, AnsHost, ManagerQueryFns, VersionControl,
+    Abstract, AbstractAccount, AccountFactoryQueryFns, AnsHost, ManagerQueryFns, RegisteredModule,
+    VersionControl,
 };
 use cosmwasm_std::{Addr, BlockInfo, Coin, Empty, Uint128};
 use cw_orch::prelude::*;
@@ -280,6 +281,26 @@ impl<Chain: CwEnv> AbstractClient<Chain> {
     pub fn next_local_account_id(&self) -> AbstractClientResult<u32> {
         let sequence = self.abstr.account_factory.config()?.local_account_sequence;
         Ok(sequence)
+    }
+
+    /// Get address of instantiate2 module
+    /// If used for upcoming account this supposed to be used in pair with [`AbstractClient::next_local_account_id`]
+    pub fn module_instantiate2_address<M: RegisteredModule>(
+        &self,
+        account_id: &AccountId,
+    ) -> AbstractClientResult<Addr> {
+        let salt = generate_instantiate_salt(account_id);
+        let wasm_querier = self.environment().wasm_querier();
+        let creator = self.abstr.module_factory.addr_str()?;
+        let code_id = self.version_control().get_module_code_id(
+            M::module_id(),
+            abstract_core::objects::module::ModuleVersion::Version(M::module_version().to_owned()),
+        )?;
+
+        let addr = wasm_querier
+            .instantiate2_addr(code_id, creator, salt)
+            .map_err(Into::into)?;
+        Ok(Addr::unchecked(addr))
     }
 }