+++ framework/CHANGELOG.md
@@ -21,12 +21,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `unchecked_account_id` method on version control.
 - Ability to provide expected local AccountId
 - Reinstallation of the same version of an app is now disabled
+- `.authorize_on_adapters` method on `Application` for authorizing application on adapters
 - Added method to assign expected `.account_id` for Abstract Client Account builder
 - `.next_local_account_id` for `AbstractClient` to query next local account sequence
+- `.module_instantiate2_address` for `AbstractClient` to get predicted address
 
 ### Changed
 
 - Updated UsageFee api to use `Address`, instead of `Api` + unchecked address
+- Tests now use `MockBech32` due to use of instantiate2.
 
 ### Removed